--- a/src/builtin/tags/render.ts
+++ b/src/builtin/tags/render.ts
@@ -64,12 +64,12 @@ export default {
       scope['forloop'] = new ForloopDrop(collection.length)
       for (const item of collection) {
         scope[alias] = item
-        const templates = yield liquid._parseFile(filepath, childCtx.opts, childCtx.sync)
+        const templates = yield liquid.parseFileImpl(filepath, childCtx.sync)
         yield renderer.renderTemplates(templates, childCtx, emitter)
         scope.forloop.next()
       }
     } else {
-      const templates = yield liquid._parseFile(filepath, childCtx.opts, childCtx.sync)
+      const templates = yield liquid.parseFileImpl(filepath, childCtx.sync)
       yield renderer.renderTemplates(templates, childCtx, emitter)
     }
   }