--- a/src/builtin/tags/render.ts
+++ b/src/builtin/tags/render.ts
@@ -1,16 +1,16 @@
 import { assert } from '../../util/assert'
 import { ForloopDrop } from '../../drop/forloop-drop'
 import { toEnumerable } from '../../util/collection'
-import { evalQuotedToken, TypeGuards, Tokenizer, evalToken, Hash, Emitter, TagToken, Context, TagImplOptions } from '../../types'
+import { Liquid } from '../../liquid'
+import { Token, Template, evalQuotedToken, TypeGuards, Tokenizer, evalToken, Hash, Emitter, TagToken, Context, TagImplOptions } from '../../types'
 
 export default {
+  parseFilePath,
+  renderFilePath,
   parse: function (token: TagToken) {
     const args = token.args
     const tokenizer = new Tokenizer(args, this.liquid.options.operatorsTrie)
-    this.file = this.liquid.options.dynamicPartials
-      ? tokenizer.readValue()
-      : tokenizer.readFileName()
-    assert(this.file, () => `illegal argument "${token.args}"`)
+    this['file'] = this.parseFilePath(tokenizer, this.liquid)
 
     while (!tokenizer.end()) {
       tokenizer.skipBlank()
@@ -40,14 +40,9 @@ export default {
     this.hash = new Hash(tokenizer.remaining())
   },
   render: function * (ctx: Context, emitter: Emitter) {
-    const { liquid, file, hash } = this
-    const { renderer } = liquid
-    const filepath = ctx.opts.dynamicPartials
-      ? (TypeGuards.isQuotedToken(file)
-        ? yield renderer.renderTemplates(liquid.parse(evalQuotedToken(file)), ctx)
-        : evalToken(file, ctx))
-      : file.getText()
-    assert(filepath, () => `illegal filename "${file.getText()}":"${filepath}"`)
+    const { liquid, hash } = this
+    const filepath = yield this.renderFilePath(this['file'], ctx, liquid)
+    assert(filepath, () => `illegal filename "${filepath}"`)
 
     const childCtx = new Context({}, ctx.opts, ctx.sync)
     const scope = yield hash.render(ctx)
@@ -65,12 +60,47 @@ export default {
       for (const item of collection) {
         scope[alias] = item
         const templates = yield liquid.parseFileImpl(filepath, childCtx.sync)
-        yield renderer.renderTemplates(templates, childCtx, emitter)
+        yield liquid.renderer.renderTemplates(templates, childCtx, emitter)
         scope.forloop.next()
       }
     } else {
       const templates = yield liquid.parseFileImpl(filepath, childCtx.sync)
-      yield renderer.renderTemplates(templates, childCtx, emitter)
+      yield liquid.renderer.renderTemplates(templates, childCtx, emitter)
     }
   }
 } as TagImplOptions
+
+type ParsedFileName = Template[] | Token | string | undefined
+
+/**
+ * @return null for "none",
+ * @return Template[] for quoted with tags and/or filters
+ * @return Token for expression (not quoted)
+ * @throws TypeError if cannot read next token
+ */
+export function parseFilePath (tokenizer: Tokenizer, liquid: Liquid): ParsedFileName | null {
+  if (liquid.options.dynamicPartials) {
+    const file = tokenizer.readValue()
+    if (file === undefined) throw new TypeError(`illegal argument "${tokenizer.input}"`)
+    if (file.getText() === 'none') return null
+    // for filenames like "files/{{file}}", eval as liquid template
+    if (TypeGuards.isQuotedToken(file)) {
+      const tpls = liquid.parse(evalQuotedToken(file))
+      // for filenames like "files/file.liquid", extract the string directly
+      if (tpls.length === 1) {
+        const first = tpls[0]
+        if (TypeGuards.isHTMLToken(first)) return first.getText()
+      }
+      return tpls
+    }
+    return file
+  }
+  const filepath = tokenizer.readFileName().getText()
+  return filepath === 'none' ? null : filepath
+}
+
+export function renderFilePath (file: ParsedFileName, ctx: Context, liquid: Liquid) {
+  if (typeof file === 'string') return file
+  if (Array.isArray(file)) return liquid.renderer.renderTemplates(file, ctx)
+  return evalToken(file, ctx)
+}