--- a/test/unit/tokens/property-access-token.ts
+++ b/test/unit/tokens/property-access-token.ts
@@ -8,14 +8,14 @@ chai.use(sinonChai)
 const expect = chai.expect
 
 describe('PropertyAccessToken', function () {
-  describe('getVariableAsText', function () {
+  describe('#propertyName', function () {
     it('should return correct value for IdentifierToken', function () {
       const token = new PropertyAccessToken(new IdentifierToken('foo', 0, 3), [], 3)
-      expect(token.getVariableAsText()).to.equal('foo')
+      expect(token.propertyName).to.equal('foo')
     })
     it('should return correct value for QuotedToken', function () {
       const token = new PropertyAccessToken(new QuotedToken('"foo bar"', 0, 9), [], 9)
-      expect(token.getVariableAsText()).to.equal('foo bar')
+      expect(token.propertyName).to.equal('foo bar')
     })
   })
 })