… test/integration/liquid/strict.ts