* parse filenames in parse() instead of render() if applicable
* postpone `resolve()` for partial fallback
* use file specifier instead of resolved filepath as cache key

Before this update (50x slower than Handlebars, even slower than React):

```
➜  benchmark git:(refs/heads/master) npm run engines
> benchmark@1.0.0 engines
> node -e 'require("./cross-engines").crossEngines()'
liquid x 788 ops/sec ±1.94% (89 runs sampled)
handlebars x 36,930 ops/sec ±1.07% (91 runs sampled)
react x 4,449 ops/sec ±2.37% (86 runs sampled)
swig x 53,257 ops/sec ±0.77% (92 runs sampled)
```

After this update (3x slower than Handlebars, 3x faster than React):

```
➜  benchmark git:(refs/heads/master) npm run engines
> benchmark@1.0.0 engines
> node -e 'require("./cross-engines").crossEngines()'
liquid x 12,311 ops/sec ±1.27% (90 runs sampled)
handlebars x 36,997 ops/sec ±1.31% (87 runs sampled)
react x 4,418 ops/sec ±3.08% (87 runs sampled)
swig x 50,481 ops/sec ±2.81% (87 runs sampled)
```